fix: move code to communication.py and rename function
